# Speed up Helpers::Pair test